+++ trycourier/client.py
@@ -1,12 +1,13 @@
 from os import environ
 
+from .automations import Automations
 from .exceptions import CourierAPIException
 from .session import CourierAPISession
 from .lists import Lists
 from .messages import Messages
 from .profiles import Profiles
 
-__version__ = '1.7.0'
+__version__ = '1.8.0'
 
 
 class Courier(object):
@@ -53,6 +54,7 @@ class Courier(object):
             password = environ.get('COURIER_AUTH_PASSWORD', None)
             self.session.init_basic_auth(username, password)
 
+        self.automations = Automations(self.base_url, self.session)
         self.lists = Lists(self.base_url, self.session)
         self.messages = Messages(self.base_url, self.session)
         self.profiles = Profiles(self.base_url, self.session)